stock/refactoring/move-component-files
