+++ libs/lib-stock/src/index.ts
@@ -1,13 +1,13 @@
 export * from './lib/forms/delete-user-form/delete-user-form';
 export * from './lib/forms/change-password-form/change-password-form';
-export * from './lib/user-profile/user-profile';
+export * from './lib/templates/user-profile/user-profile';
 export * from './lib/forms/register-form/register-form';
-export * from './lib/admin-page/admin-page';
-export * from './lib/main-page/main-page';
+export * from './lib/pages/admin-page/admin-page';
+export * from './lib/pages/main-page/main-page';
 export * from './lib/forms/login-form/login-form';
-export * from './lib/homepage/homepage';
+export * from './lib/pages/homepage/homepage';
 export * from './lib/header/header';
 export * from './lib/app-root/app-root';
-export * from './lib/dashboard/dashboard';
+export * from './lib/templates/dashboard/dashboard';
 export * from './lib/menu/menu';
-export * from './lib/companies/companies';
+export * from './lib/templates/companies/companies';
