+++ libs/lib-stock/src/lib/pages/admin-page/admin-page.tsx
