+++ libs/lib-stock/src/lib/pages/homepage/homepage.spec.tsx
