+++ libs/lib-stock/src/lib/pages/homepage/homepage.tsx
