+++ libs/lib-stock/src/lib/pages/main-page/main-page.spec.tsx
