+++ libs/lib-stock/src/lib/templates/companies/companies.spec.tsx
