+++ libs/lib-stock/src/lib/templates/companies/constatnts.ts
