+++ libs/lib-stock/src/lib/templates/dashboard/dashboard.spec.tsx
