+++ libs/lib-stock/src/lib/templates/user-profile/user-profile.spec.tsx
