+++ libs/ui/src/index.ts
@@ -1,7 +1,7 @@
 export * from './lib/molecules/table-content/table-content';
 export * from './lib/molecules/table-header/table-header';
 export * from './lib/atoms/table-header-cell/table-header-cell';
-export * from './lib/molecules/table/table';
+export * from './lib/organisms/table/table';
 export * from './lib/atoms/table-row/table-row';
 export * from './lib/atoms/error-info/error-info';
 export * from './lib/atoms/loader/loader';
